Improve portability of the curvefit test (#761)
* DRR - Cpptraj: Compare results from more stable curve fit

* DRR - Cpptraj: The multi-exponential fit with penalties is not very numerically stable, so do not check the exact results since exponential ordering can change without changing the quality of the resulting fit. Just check the curve. Also now check absolute tolerances of the fit curves; allow change of 1 unit in the last decimal place.

* DRR - Cpptraj: Change test target to test.showerrors to provide more info on failing tests with Jenkins.

* DRR - Cpptraj: Clean up the OSX build on travis. homebrew/science has been deprecated for a while, and arpack is now bundled with cpptraj.

* DRR - Cpptraj: Fix error string. Make BZerror static

* DRR - Cpptraj: Improve bzip2 error messages and ensure they are written to stderr

* DRR - Cpptraj: Keep track of EOF status when reading bzip2 files since calling bzRead after BZ_STREAM_END is considered a sequence error. When writing bzip2 files, any status but BZ_OK should be an error.

* DRR - Cpptraj: Hide debug info

* DRR - Cpptraj: Revision bump for bzip fixes

* DRR - Cpptraj: Store distance for representative frame as double; no need to take the precision loss and it may improve the stability of tests.

* DRR - Cpptraj: Fix potential issue where MPI send/receive types didnt match. Ensure master resizes the format based on max size found in rank.

* DRR - Cpptraj: For parallel, rank print a data set not found error so its clear the error is different data on different threads.

* DRR - Cpptraj: When a data set is not present on a rank, skip the calculation on that rank and warn.

+++ src/ClusterList.cpp
@@ -109,7 +109,7 @@ void SetBestRepFrame(ClusterNode& node, RepMap const& reps) {
   if (!reps.empty()) {
     node.BestReps().clear();
     for (RepMap::const_iterator it = reps.begin(); it != reps.end(); ++it) {
-      node.BestReps().push_back( ClusterNode::RepPair(it->second, (float)it->first) );
+      node.BestReps().push_back( ClusterNode::RepPair(it->second, it->first) );
     }
   }
 }
